requirements.txt
pyyaml>=6.0
trimesh>=3.23.0
pytorch3d>=0.7.4
opencv-python>=4.8.0
tqdm>=4.66.0
pillow>=10.0.0
scipy>=1.11.0
matplotlib>=3.7.0
bpy>=1.0.0  # Optional, for Blender API


anndata==0.12.10
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array-api-compat==1.14.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
Brotli==1.2.0
cached-property==1.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
donfig==0.8.1.post1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
geojson==3.2.0
google-crc32c==1.8.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hopcroftkarp==1.2.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter_client==8.8.0
jupyter-console==6.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
legacy-api-wrap==1.5
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
miniball==1.2.0
mistune==3.2.0
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
mpi4py==4.0.3
mpmath==1.3.0
muspan==1.2.2
natsort==8.4.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.3
notebook_shim==0.2.4
numcodecs==0.16.5
numexpr==2.14.1
numpy==2.4.2
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.0
parso==0.8.6
persim==0.3.8
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.1
POT==0.9.6.post1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.0
pycparser==2.22
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-ranger==0.1.1
pytz==2026.1.post1
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
ripser==0.6.14
rpds-py==0.30.0
rust-geo-python==0.1.6
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==80.10.2
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tables==3.10.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.10.0+cpu
torch-optimizer==0.3.0
torchaudio==2.10.0+cpu
torchdiffeq==0.2.5
torchvision==0.25.0+cpu
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
typing_utils==0.1.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
voila==0.5.11
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
widgetsnbextension==4.0.15
win_inet_pton==1.1.0
wrapt==2.1.1
zarr==3.1.5
zipp==3.23.0
